Add multi year support (#405)
* Update footer year

* Extract 2024 to it's own year path

Add redirect pages

* Add year to results page

* Create catchall redirect page

* Fix landing page links

* Add year selector to footer

* Add 2025 landing page

* Add sitemap

* Add year to actions

* Format

* Add missing year from get results

* Fix years

* bump deps

* Fix uploadthing CSP error

* Fix some errors

* Use SimpleLogin provider from auth.js

* Enhance file mime type zod schema

* Add 2025 results

* Fix eslint

* Refactor

* Refactors

* Bump TS

* Eslint stuff

* Eslint fix

Add year to manifest

* Fix sitemap

* Fix active link

+++ src/app/lib/actions.ts
@@ -1,44 +1,31 @@
 'use server';
 
 import { revalidatePath } from 'next/cache';
-import { deleteResultByKey, getResultItems, insertResult, isAdmin, updateAvatar } from '../db/data';
+import { deleteResultByKey, getResultItems, isAdmin } from '../db/data';
 import { auth } from '@/app/lib/auth';
 import { isNullOrEmpty } from '@/app/utils';
-import { type Result, type ResultItem, type ResultType } from '@/app/lib/types';
+import { type ResultItem } from '@/app/lib/types';
 import { env } from '@/app/lib/env';
 
-export async function deleteResult(key: string) {
+export async function deleteResult(key: string, year: number) {
+  if (!(await canEditResults({ year }))) {
+    return {
+      error: 'Hiba történt a törlés során! Nincs jogosultság a törléshez.',
+    };
+  }
   await deleteResultByKey(key);
-  revalidatePath('/eredmenyek');
+  revalidatePath(`/${year.toString()}/eredmenyek`);
+  return null;
 }
 
-export async function uploadResult(result: { key: string; result: Result; type: ResultType }) {
-  await insertResult({ ...result, year: 0 });
-  revalidatePath('/eredmenyek');
-}
-
-export async function canEditResults() {
+export async function canEditResults({ year }: Readonly<{ year: number }>) {
   const session = await auth();
   const email = session?.user?.email;
-  return !isNullOrEmpty(email) && (await isAdmin(email));
+  return !isNullOrEmpty(email) && (await isAdmin(email, year));
 }
 
-export async function updateUserAvatar({
-  id,
-  avatar,
-}: Readonly<{
-  id: string | null | undefined;
-  avatar: string | null;
-  provider: string | null | undefined;
-}>) {
-  if (isNullOrEmpty(id)) {
-    return;
-  }
-  return await updateAvatar({ id, avatar: avatar ?? null });
-}
-
-export async function getResults(sportag: string): Promise<ResultItem[]> {
-  const results = await getResultItems(sportag);
+export async function getResults(sportag: string, year: number): Promise<ResultItem[]> {
+  const results = await getResultItems(sportag, year);
   for (const result of results) {
     result.url = `https://${env.UPLOADTHING_APP_ID}.ufs.sh/f/${result.key}`;
   }
